+++ system/src/hooks.mjs
@@ -3,6 +3,7 @@ import { ChatMessageHooks } from "./hooks/chat-messages.mjs";
 import { DropLightsourceHooks } from "./hooks/drop-lightsource-on-scene.mjs";
 import { EffectHooks } from "./hooks/effects.mjs";
 import { EffectPanelHooks } from "./hooks/effect-panel.mjs";
+import { ForeignDocumentHooks } from "./hooks/ForeignDocumentHooks.mjs";
 import { LightSourceTrackerHooks } from "./hooks/light-source-tracker.mjs";
 import { NPCHooks } from "./hooks/npc.mjs";
 import { ShadowdarklingImport } from "./hooks/shadowdarkling-import.mjs";
@@ -14,6 +15,7 @@ export const HooksSD = {
 			ChatMessageHooks,
 			DropLightsourceHooks,
 			EffectHooks,
+			ForeignDocumentHooks,
 			LightSourceTrackerHooks,
 			NPCHooks,
 		];