+++ system/src/sheets/ActorSheetSD.mjs
@@ -10,6 +10,10 @@ export default class ActorSheetSD extends ActorSheet {
 			event => this._onRollHP(event)
 		);
 
+		html.find(".pc-roll-initiative .rollable").click(
+			event => this._onRollInitiative(event)
+		);
+
 		html.find(".open-item").click(
 			event => this._onOpenItem(event)
 		);
@@ -127,15 +131,15 @@ export default class ActorSheetSD extends ActorSheet {
 	}
 
 	_getItemContextOptions() {
-		const canEdit = function(element) {
+		const canEdit = function(element, actor) {
 			let result = false;
 			const itemId = element.data("item-id");
 
 			if (game.user.isGM) {
 				result = true;
 			}
-			else {
-				result = game.user.character.items.find(item => item._id === itemId)
+			else if (actor.canUserModify(game.user)) {
+				result = actor.items.find(item => item._id === itemId)
 					? true
 					: false;
 			}
@@ -147,7 +151,7 @@ export default class ActorSheetSD extends ActorSheet {
 			{
 				name: game.i18n.localize("SHADOWDARK.sheet.general.item_edit.title"),
 				icon: '<i class="fas fa-edit"></i>',
-				condition: element => canEdit(element),
+				condition: element => canEdit(element, this.actor),
 				callback: element => {
 					const itemId = element.data("item-id");
 					const item = this.actor.items.get(itemId);
@@ -157,7 +161,7 @@ export default class ActorSheetSD extends ActorSheet {
 			{
 				name: game.i18n.localize("SHADOWDARK.sheet.general.item_delete.title"),
 				icon: '<i class="fas fa-trash"></i>',
-				condition: element => canEdit(element),
+				condition: element => canEdit(element, this.actor),
 				callback: element => {
 					const itemId = element.data("item-id");
 					this._onItemDelete(itemId);
@@ -217,6 +221,13 @@ export default class ActorSheetSD extends ActorSheet {
 		event.preventDefault();
 
 		this.actor.rollHP();
+	}
+
+	async _onRollInitiative(event) {
+		event.preventDefault();
+
+		// User the default roll available to each Actor / Token
+		await this.actor.rollInitiative({ createCombatants: true, rerollInitiative: true});
 	}
 
 	async _onRollAbilityCheck(event) {