+++ RELEASE_NOTES.md
@@ -1,3 +1,18 @@
+## v.1.3.5
+
+We welcome @gatesvp as a contributor to the system!
+
+### Bugfixes
+* [#452] Refers the issue tracker to the actual issue tracker instead of repo
+* [#450] Rolling NPC HP now correctly adds CON mod bonus (minimum 1) to HP
+* [#461] Allows players with edit permission to get context menu on items on a character to edit and delete
+* [#462] Properly awaiting Active Effects to be created for V11 compatability
+
+### Enhancements
+* [#452] Added special attack to chat card when applicable
+* [#277] Partial implementation provides quick way to apply damage and healing directly from the chat card
+* [#311] Initiative field on PC sheets
+
 ## v.1.3.4
 
 ### Enhancements